fix the link to a section within the org file
Unfortunately, the link to a section does not work anymore, with the change I made it will still work on GitHub, although not in Emacs. If the heading name is more than a single word, they need to be attached by hyphen, e.g. if the heading is "two word section" the link should be "#two-word-section" . I hope you find it helpful.